Make implicit type conversion in DVB CA code explicit for C++11 compliance.

+++ mythtv/libs/libmythtv/recorders/dvbdev/dvbci.cpp
@@ -1102,7 +1102,13 @@ bool cCiDateTime::SendDateTime(void)
      int MJD = 14956 + D + int((Y - L) * 365.25) + int((M + 1 + L * 12) * 30.6001);
 #define DEC2BCD(d) (uint8_t(((d / 10) << 4) + (d % 10)))
      struct tTime { unsigned short mjd; uint8_t h, m, s; short offset; };
-     tTime T = { mjd : htons(MJD), h : DEC2BCD(tm_gmt.tm_hour), m : DEC2BCD(tm_gmt.tm_min), s : DEC2BCD(tm_gmt.tm_sec), offset : htons(tm_loc.tm_gmtoff / 60) };
+     tTime T = {
+         mjd : htons(MJD),
+         h : DEC2BCD(tm_gmt.tm_hour),
+         m : DEC2BCD(tm_gmt.tm_min),
+         s : DEC2BCD(tm_gmt.tm_sec),
+         offset : static_cast<short>(htons(tm_loc.tm_gmtoff / 60))
+     };
      dbgprotocol("%d: ==> Date Time\n", SessionId());
      SendData(AOT_DATE_TIME, 7, (uint8_t*)&T);
      //XXX return value of all SendData() calls???
